requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.3b0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycryptodome==3.9.0
toml==0.10.0
Werkzeug==0.16.0
